+++ MapleServer2/AIScripts/AiHelper.cs
@@ -1,16 +1,23 @@
-﻿using MapleServer2.AIScripts._29000282;
-using MapleServer2.Managers.Actors;
-using MapleServer2.Types;
+﻿using MapleServer2.Types;
 
 namespace MapleServer2.AIScripts;
 
 public static class AiHelper
 {
-    public static readonly Dictionary<int, AIState> AIStates = new()
+    public static readonly Dictionary<string, AIState> AIStates = new()
     {
         {
-            29000282,
-            new BelmaBehavior()
+            "AI_DefaultNew.xml",
+            new DefaultNew()
+        },
+        {
+            "AI_PrisonGovernor2_battle.xml",
+            new PrisonGovernor2Battle()
         }
     };
+
+    public static AIState GetAIState(string aiName)
+    {
+        return AIStates.ContainsKey(aiName) ? AIStates[aiName] : new DefaultNew();
+    }
 }
